# rest/config/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Liverpool FC API",
    "version": "1.0.0",
    "description": "REST API for managing Liverpool Football Club data including players, matches, and more"
  },
  "servers": [
    {
      "url": "http://localhost:8989/api/v1"
    }
  ],
  "paths": {
    "/competitions": {
      "get": {
        "summary": "Get all competitions",
        "tags": [
          "Competitions"
        ],
        "responses": {
          "200": {
            "description": "List of competitions",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of competitions",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Competition"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No competitions found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create new competition",
        "tags": [
          "Competitions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Competition created",
            "headers": {
              "Location": {
                "description": "URL of created competition",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/CompetitionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/competitions/{id}": {
      "get": {
        "summary": "Get competition by ID",
        "tags": [
          "Competitions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Competition ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Competition found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/CompetitionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Competition not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update competition",
        "tags": [
          "Competitions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Competition ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Competition updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/CompetitionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Competition not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Partially update competition",
        "tags": [
          "Competitions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Competition ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "yearOfCreation": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Competition updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/CompetitionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Competition not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete competition",
        "tags": [
          "Competitions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Competition ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Competition deleted"
          },
          "404": {
            "description": "Competition not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Competition": {
        "type": "object",
        "required": [
          "name",
          "type",
          "yearOfCreation"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Competition name (must be unique)"
          },
          "type": {
            "type": "string",
            "enum": [
              "LEAGUE",
              "CUP",
              "FRIENDLY"
            ],
            "description": "Competition type"
          },
          "yearOfCreation": {
            "type": "number",
            "minimum": 1800,
            "description": "Year when competition was created (cannot be in future)"
          }
        }
      },
...
    },
    "responses": {
      "CompetitionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Competition"
          },
          "_links": {
            "type": "object",
            "properties": {
              "self": {
                "type": "string"
              },
              "collection": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
}